Pull useSqlEditor to a separate hook
